Remove requirements for Upgrade and Connection headers in Websocket handshakes
We have a particular client that is behind a Blue Coat proxy, and their Websocket handshakes are missing the Upgrade header and have an incorrect Connection header.  Here is what these handshakes look like:

https://gist.github.com/1573310

Compare to what a protocol-conforming handshake should look like:

https://gist.github.com/1573314

Because we need to support this particular client we are removing the checks for these headers.  What will be the side effects of this?